+++ changelog.md
@@ -11,6 +11,8 @@ See also the commit log at github: https://github.com/Arduino-IRremote/Arduino-I
 - Fix for #1028 - Prevent long delay caused by overflow when frame duration < repeat period - Thanks to Stephen Humphries!
 - Support for ATtiny816 - Thanks to elockman.
 - Added Bang&Olufsen protocol. #1030.
+- 3. parameter of function "void begin(uint_fast8_t aSendPin, bool aEnableLEDFeedback, uint_fast8_t aFeedbackLEDPin)" is not optional anymore
+     and function is now only available if IR_SEND_PIN is not defined. #1033.
 
 ## 3.9.0
 - Improved documentation with the help of [ElectronicsArchiver}(https://github.com/ElectronicsArchiver).